requirements.txt
tqdm==4.24.0
image==1.5.24
pyyaml==3.13
